Cleaned up obsolete debug code

+++ domain/src/main/kotlin/com/seanshubin/jvmspec/domain/api/ApiConstant.kt
@@ -3,14 +3,16 @@ package com.seanshubin.jvmspec.domain.api
 import com.seanshubin.jvmspec.domain.primitive.ConstantPoolTag
 
 sealed interface ApiConstant {
-    data class Constant(val index: Int, val tagId: Int, val tagName: String, val parts: List<ApiConstant>) : ApiConstant
+    data class Constant(val index: UShort, val tagId: UByte, val tagName: String, val parts: List<ApiConstant>) :
+        ApiConstant
+
     data class StringValue(val s: String) : ApiConstant
     data class IntegerValue(val i: Int) : ApiConstant
     data class FloatValue(val f: Float) : ApiConstant
     data class LongValue(val l: Long) : ApiConstant
     data class DoubleValue(val d: Double) : ApiConstant
-    data class ReferenceKindValue(val id: Int, val name: String) : ApiConstant
-    data class IndexValue(val index: Int) : ApiConstant
+    data class ReferenceKindValue(val id: UByte, val name: String) : ApiConstant
+    data class IndexValue(val index: UShort) : ApiConstant
     companion object {
         fun utf8ToString(constant: Constant): String {
             require(constant.tagName == ConstantPoolTag.UTF8.name)
